+++ packages/beff-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @beff/cli
 
+## 0.0.33
+
+### Patch Changes
+
+- fix export declare bug
+
 ## 0.0.32
 
 ### Patch Changes