+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @beff/client
 
+## 0.0.33
+
+### Patch Changes
+
+- fix export declare bug
+- Updated dependencies
+  - @beff/cli@0.0.33
+
 ## 0.0.32
 
 ### Patch Changes